Driver class -Update-
- I found a way to read the file without loops and stored it in a stream
- Now working on sorting and displaying the Shapes.
