+++ inputShapes.txt
@@ -1,0 +1,10 @@
+Rectangle,2.5,4
+Circle,3
+Circle,6
+Rectangle,5,5
+Rectangle,1,4
+Rectangle,8,9.5
+Circle,4.5
+Circle,9
+Circle,1
+Rectangle,7.5,1.5
